Move entity schema and service into building package
