+++ .gitignore
@@ -339,7 +339,7 @@ ASALocalRun/
 
 # Sqlite 3 database files
 *.sqlite3
-back/pkg/entityservice/debug.test
+back/pkg/building/db/entityservice/debug.test
 back/pkg/building/room/models/tmp/test.xml
 back/pkg/building/room/models/tmp/test.json
 .vscode/studiguide.code-workspace
