+++ back/cmd/studi-guide/main.go
@@ -4,9 +4,9 @@ import (
 	"log"
 	"studi-guide/cmd/studi-guide/server"
 	"studi-guide/docs"
+	"studi-guide/pkg/building/db/entityservice"
 	"studi-guide/pkg/building/location"
 	"studi-guide/pkg/building/room/models"
-	"studi-guide/pkg/entityservice"
 	"studi-guide/pkg/env"
 	"studi-guide/pkg/navigation"
 	"studi-guide/pkg/navigation/services"
