+++ back/cmd/studi-guide/server/studiguide_server.go
@@ -10,11 +10,11 @@ import (
 	"net/http"
 	"net/url"
 	"os"
+	"studi-guide/pkg/building/db/entityservice"
 	"studi-guide/pkg/building/info"
 	"studi-guide/pkg/building/location"
 	maps "studi-guide/pkg/building/map"
 	"studi-guide/pkg/building/room/controllers"
-	"studi-guide/pkg/entityservice"
 	"studi-guide/pkg/env"
 	navigation "studi-guide/pkg/navigation/controllers"
 	"studi-guide/pkg/navigation/services"
