+++ back/pkg/building/db/ent/schema/mapitem.go
