+++ back/pkg/building/db/ent/schema/pathnodegroup.go
