+++ back/pkg/building/room/controllers/roomcontroller.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"github.com/gin-gonic/gin"
 	"net/http"
+	"studi-guide/pkg/building/db/entityservice"
 	"studi-guide/pkg/building/room/models"
-	"studi-guide/pkg/entityservice"
 )
 
 type RoomController struct {
