+++ back/pkg/building/room/models/ImportMapItems.go
@@ -1,7 +1,7 @@
 package models
 
 import (
-	"studi-guide/pkg/entityservice"
+	"studi-guide/pkg/building/db/entityservice"
 	"studi-guide/pkg/navigation"
 )
